Moved non-view components into common folder
