+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import "./common/Button.css"
 import {responsiveFontSizes} from "@material-ui/core";
 import {createMuiTheme, StylesProvider, ThemeProvider} from "@material-ui/core/styles";
-import {ProvideAuth} from "./Auth/Auth";
+import {ProvideAuth} from "./common/Auth";
 import BaseRouter from "./BaseRouter";
