+++ src/common/Managers/CookieManager.js
