+++ src/common/Managers/EndpointManager.js
